cryptohome: Add a new parameter to DeriveCallback
Adds a new parameter into DeriveCallback that allows for the reporting
of non-failure problems. In particular this can be used to report that
an auth block is valid, but obsolete and needs to be re-created.

Also adds in an implementation for this for the pinweaver auth block,
which will (if you have pin migration enabled) report the credential as
obsolete if it doesn't have a gradual backoff delay schedule.

I took a shot at adding this as a separate standalone function, but that
turned out to be rather awkward to implement because each block then had
to duplicate some of the Derive code to actually make sure that the
state you passed it made sense. While this requires an annoyingly larger
delta to add a parameter all over the place, it makes the implementation
(where most derive functions report nothing) much simpler.

BUG=b:272560921
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I8fdf0708903bab1ab5acfa6124d7bcf18b0c5584
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4401730
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>