shill: Accept prefixlen == 0 in shill client
Prefix length == 0 usually means there will be some problems with the
network connectivity, but it's kind of a valid value (e.g., one can use
`ip addr add` to configure an address with `/0` on an interface). Since
we are not enforcing this restriction (i.e., prefix length should not be
0) inside, it's better to avoid rejecting it in shill clients (in both
patchpanel and shill/dbus/client).

BUG=b:276849438
TEST=FEATURES="test" emerge-${BOARD} shill-dbus-client patchpanel

Change-Id: I6d7095f5bc89e472f5331f14ce5bf69ccc78c546
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4403010
Reviewed-by: Garrick Evans <garrick@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>