shill: Report metrics when connect is cancelled
Previously, metrics would not be reported if a connect request
was queued. A Connect call is queued if it is made while
the modem was not registered. This queued call is cancelled if the
modem doesn't register within a timeout, and no metrics are
emitted.

BUG=b:272820294
TEST=emerge shill; cros deploy ip shill; Connect on a bad SIM

Change-Id: I12d821f2676ad1c3e32dede1871fbf8f13f66cdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4404430
Commit-Queue: Pavan Holla <pholla@google.com>
Tested-by: Pavan Holla <pholla@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>