login: Add histogram for the state of the device policy
We are interested specifically in the cases when the enrollment is lost.
That is install attributes is present while the policy key or device
policy fail to load.

The plan is to compare this histogram to what we get from the Chrome
side in "EnrolledPolicyHasDMToken" histogram.

BUG=b:231677312
TEST=Manual test on device

Cq-Depend:4444173
Change-Id: I560f3d0c912b8e94391e7ecde5d1f8be312657b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4403582
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Igor <igorcov@chromium.org>