cryptohome: merge some UserDataAuth init functions
This merges the Initialize and PostDBusInitialize operations in
UserDataAuth, as well as folding the bus setting functions in as well by
making those buses parameters.

This shouldn't have an impact on startup time, as all of these
operations were happening before starting anyway and so we're not moving
any post-serving init to be pre-serving.

Unfortunately this doesn't immediately help as much as I had thought it
might because I now realize that a bigger part of the init headache is
stuff initialized on the origin thread vs the mount thread, but I still
think that can be addressed with some more work and this merge will make
that easier.

BUG=b:276453357
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I589b929747057b0219a157b117a0248e8a1cad58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4406834
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>