shill: tethering: add tethering stop timer
Tethering stop process is event driven and could be trapped in a
dangling state when something wrong with the corresponding component and
the driving event is not received.

Add a tethering stop timer and handler to handle this case and reset the
state to normal. Post dbus response when necessary.

BUG=b:277276931
TEST=FEATURES=test emerge-${BOARD} shill
TEST=on brya, check if stopping tethering returns correct result.
TEST=on brya, hack the code to miss the OnUpstreamNetworkReleased event,
check if tethering state has been reset to idle and error message is
returned.

Change-Id: Iab35a7d08dd65344f0c2e9049855bfcf2899f27c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4409185
Commit-Queue: Jintao Lin <jintaolin@chromium.org>
Tested-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>