lorgnette: Separate d-bus interface from manager class
The Manager class is both the d-bus implementation and a ton of
supporting code.  It has gotten fairly unwieldy and is only going to get
worse when we add a bunch of new d-bus entry points to support advanced
scanning.

Instead of using Manager as both the d-bus adaptor and the
implementation, separate out the adaptor into a new class.  This doesn't
immediately buy us any code savings, but it will allow future changes to
be brought in without continuing to bloat up the existing class.  It
also makes it easy to tell where all the d-bus entry points are, since
they're no longer mixed with other implementation code.

No direct scanning changes in this CL, but it does introduce a new scope
logging class so that logging d-bus calls is handled consistently.

BUG=b:276909622
TEST=tast run '("paper-io_scanning")'

Change-Id: Id685da5e7e047f8c3f1e29909380dfe55e6744b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4409083
Reviewed-by: Paul Moy <pmoy@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>