dns-proxy: Avoid starting child processes when shut down
There is a somewhat race that happens on (a) DNS proxy being stopped and
(b) Chrome flag update / user login. The behavior resulted in the
controller spawning its child processes before stopping itself. This
causes multiple DNS proxy instances to run in the lab.

This should not impact users' devices as it is not expected for DNS
proxy to be stopped.

BUG=None
TEST=FEATURES=test emerge-hatch-arc-t dns-proxy
TEST=CQ

Change-Id: Icb7620ee0534e7a8966f7d07bc0deb1a60c24ae2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4402488
Reviewed-by: Garrick Evans <garrick@chromium.org>
Tested-by: Jason Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Iman <jasongustaman@chromium.org>