R container: mount /dev/{uinput,uhid,input}
These files are required by R CTS tests, but were not required in P.
Create them for the guest. This will be a no-op for ARCVM, which already
enables and isolates activity of these files from the host.

For R Container we expose the host's uinput and uhid files as-is to the
guest, and a subset of /dev/input/event* files. /dev/uinput and
/dev/input/event* are all character devices.

/dev/uinput and /dev/uhid are created directly in the /dev mount
(tmpfs), while /dev/input is bind-mounted to /dev/arc_input. The sets of
files in /dev/input and /dev/arc_input on the host are non-overlapping.
Usually (not R container), 99-inputcontrol.rules adds all /dev/input
entries upon creation. 95-arcinput.rules removes files from /dev/input
and adds them to arc_input when they correspond to a CTS test.

/dev/input and /dev/arc_input must be disjoint sets of devices. This is
because Chrome forwards events in the host /dev/input via Wayland to
ARC. If ARC is allowed to access the device from /dev/arc_input *and*
the events are sent to ARC by Chrome, ARC receives duplicated events. In
the case of uhid joysticks, the duplicated events are also modified in a
way that the test case does not recognize them.

95-arcinput.rules creates the files in arc_input with the correct
initial permissions, otherwise system_server will try to open the file
before it has the permission, and fail.

As an aside, the initial permission issue is handled differently in
ARCVM. The kdevtmpfs kernel task creates the file in the devtmpfs file
system with 600, but init and child processes actually mount a separate
tmpfs at /dev. Then ueventd in the guest creates the file in the tmpfs
with the correct permissions already set.

This re-works https://crrev.com/c/4549540 with these improvements:

 * ARC can still boot if the udev rules are not installed.

 * Files are in exactly one of /dev/input or /dev/arc_input, rather than
   both.

 * Expose /dev/uhid too, since it is closely-related to the added udev
   rules.

Note that we create the arc_input folder in devtmpfs rather than
/var/run/arc, as the initial devices at boot are added before /var is
mounted.

TEST=run cts -m CtsInitTestCases -t keychords#init_inotify
TEST=run cts -m CtsInitTestCases -t keychords#key
TEST=run cts -m CtsInitTestCases -t keychords#keys_in_series
TEST=run cts -m CtsInitTestCases -t keychords#keys_in_parallel
TEST=run cts -m CtsHardwareTestCases -t android.hardware.input.cts.tests.AsusGamepadTest
TEST=run cts -m CtsHardwareTestCases -t android.hardware.input.cts.tests.MicrosoftDesignerKeyboardTest
TEST=run cts -m CtsHardwareTestCases -t android.hardware.input.cts.tests.MicrosoftSculpttouchTest
TEST=run cts -m CtsHardwareTestCases -t android.hardware.input.cts.tests.MicrosoftXbox2020Test
TEST=run cts -m CtsHardwareTestCases -t android.hardware.input.cts.tests.SonyDualshock3UsbTest
TEST=remove 95-arcinput.rules, reboot, and start PlayStore app
BUG=b:277541769,b:279539420

Cq-Depend: chromium:4549800
Change-Id: I9a63c43eab26c8ea77fc048d1be80b8db1814a10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4558265
Auto-Submit: Matthew DeVore <matvore@chromium.org>
Reviewed-by: Yury Khmel <khmel@google.com>
Commit-Queue: Matthew DeVore <matvore@chromium.org>
Tested-by: Matthew DeVore <matvore@chromium.org>