R container: mirror /dev/uinput and /dev/input
These files are required by R CTS tests, but were not required in P.
Create them for the guest.

We expose the host's uinput system as-is to the guest. /dev/uinput and
/dev/input/event* are all character devices.

/dev/uinput is created directly in the /dev mount (tmpfs), which
/dev/input is bind-mounted to /dev/arc_input, which has files equivalent
to /dev/input/event* in devtmpfs on the host. /dev/arc_input is kept
up-to-date with the udev rules in this commit.

We cannot bind-mount /dev/input on the host to the guest, because it
is created with permissions 600 by kdevtmpfs, then the guest tries to
open it before udev changes permissions to 660.

This work-around is only required on container, not ARCVM. This is
because ARCVM's kdevtmpfs creates the file in the devtmpfs file system
(with 600) but init and child processes actually mount a separate tmpfs
at /dev. Then ueventd in the guest creates the file in the tmpfs with
the correct permissions already set.

Note that we create the arc_input folder in devtmpfs rather than
/var/run/arc, as the initial devices at boot are added before /var is
mounted.

TEST=run cts -m CtsInitTestCases -t keychords#init_inotify
TEST=run cts -m CtsInitTestCases -t keychords#key
TEST=run cts -m CtsInitTestCases -t keychords#keys_in_series
TEST=run cts -m CtsInitTestCases -t keychords#keys_in_parallel
TEST=stat /dev/uinput and /dev/input/* in guest vs. host
BUG=b:277541769

Change-Id: If9194b46ecc71266df3d3686b81ca511d33d0de9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4549540
Reviewed-by: Yury Khmel <khmel@google.com>
Tested-by: Matthew DeVore <matvore@chromium.org>
Commit-Queue: Matthew DeVore <matvore@chromium.org>