hiberman: Lockdown hiberimage DM devices
With CL:4348874 ("CHROMIUM: security: Block open on dm_locked
devices") the kernel does not allow DM devices with a UUID of
'dm_locked-<something>' to be opened. Set the UUID of the
DM devices that comprise the hiberimage to 'dm_locked-<name>'.

Also delete the device nodes of the DM devices as an additional
layer of security.

BUG=b:266012455
TEST=on brya:
  dmsetup info | grep dm_locked
    UUID: dm_locked-hiberimage_integrity
    UUID: dm_locked-hiberintegrity
    UUID: dm_locked-hiberimage

  no /dev/dm-N nodes for the hiberimage DMs

Change-Id: I3444d44cef0c3710aee54f696f2e24215ffd078f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4415481
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>