crash: Unclean shutdown: Log ectool uptimeinfo
When the EC reboots, it resets the AP as well. From the AP's
perspective, this is in an unclean shutdown.

To improve debuggability, when an unclean shutdown is detected, log the
output from 'ectool uptimeinfo'. This will help developers correlate EC
resets with unclean AP shutdowns and (hopefully) make it quicker/easier
to determine the source of the AP's unclean shutdown.

Sample output logged to /var/log/messages:
    2023-04-11T22:54:41.822732Z WARNING crash_reporter[4047]: Last shutdown was not clean
    2023-04-11T22:54:41.831662Z INFO crash_reporter[4047]: [ectool uptimeinfo] EC uptime: 14.803 seconds
    2023-04-11T22:54:41.831684Z INFO crash_reporter[4047]: [ectool uptimeinfo] AP resets since EC boot: 0
    2023-04-11T22:54:41.831699Z INFO crash_reporter[4047]: [ectool uptimeinfo] Most recent AP reset causes:
    2023-04-11T22:54:41.831715Z INFO crash_reporter[4047]: [ectool uptimeinfo] EC reset flags at last EC boot: soft

BUG=b:276972025
TEST=Issue 'reboot' in EC CLI, verify 'ectool uptimeinfo' output is
logged.

Change-Id: Ie7550fb18211ce4e37760669ce3d3904ea3639f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4411745
Tested-by: Tim Van Patten <timvp@google.com>
Commit-Queue: Tim Van Patten <timvp@google.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>