shill: cellular: Check that comparing APNs is consistent
Add unit tests to verify that APNs can be compared successfully after
being stored in the cache and processed by ApnList.
Enforce the list of APN properties used when comparing APNs to ensure
the comparison still works when new properties are added to MobileApn.
Log when an APN is not loaded correctly from cache.

BUG=b:277584864
TEST=Unit tests.

Change-Id: Ia89380a1aef328b4259383d849d5ad51592b63c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4412654
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>