hiberman: Return actual block device path from DeviceMapper::device_path()
Before this patch DeviceMapper::device_path() returned the path of
the symlink to the the DM device, i.e. /dev/mapper/<name>. This is
generally fine in practical terms, but not completely accurate.
Return the target of the symlink (usually /dev/dm-N) instead.

The device path can now only be determined if a DM device actually
exists, so change the return type of device_patch to a Result.

For hiberman we plan to remove the device nodes of the DM devices
that comprise the hiberimage. With that device_exists() wouldn't
return a correct result, so run 'dmsetup status' to determine
whether a device exists, instead of looking for the device nodes.
device_exists() isn't called frequently, so the overhead from
launching an external process is acceptable.

BUG=b:266012455
TEST=hibernate works on brya

Change-Id: I60b6d701e61d9e9fe25eb9b885009e5b2ad1e419
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4415480
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>