vm_tools: Wait on both event fd and signal fd during VM start
Prior to this change, there is a timed event wait during the VM start
path i.e. if we timeout waiting for the VM to start we report a failure.
However, if the crosvm process exits during this timeout we still have
to wait the entire time of the timeout. This is because the event wait
and the signal handler are both registered on the same thread. The
SIGCHLD can't be handled while we're blocking the main thread with a
timed wait.

This change solves this problem by using an event fd and epoll_waiting
on both the event fd and signal fd. This ensures we can process a
SIGCHLD event while simultaneously waiting for the VM start event fd.
Thus, we don't have to wait for the entire time.

BUG=b:260274230
TEST=Tast tests and passing crosvm with bad args and notice vmc returns
immediately.

Change-Id: I454334b87f9ae81b962292a67603de5dcb302f59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4415486
Commit-Queue: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Auto-Submit: Abhishek Bhardwaj <abhishekbh@chromium.org>