hiberman: Thicken hiberimage and hiberintergrity volumes
The creation of the device mapper devices on top of the hiberimage
and hiberintegrity thin volumes does not thicken them. When the
hibernate snapshot is created in memory towards the end of the
hibernation process both volumes are still mostly unused. The
snapshot is only written to disk after it's creation, which alters
the metadata of the thinpool if the volumes have not be thickened
beforehand. When the hibernated kernel is later resumed it still
sees the mostly unused volumes (as they were when the snapshot
image was taken), which is in mismatch with the thinpool data on
disk and can result in file system corruptions.

Thicken the hiberintegrity LV directly after it's creation (at boot
time). This can't be postponed until the system actually hibernates
because the kernel formats the LV when the dm-integrity device is
created on top of it and we don't want to overwrite the formatting.
The volume is relatively small, so thickening it on every boot
should be ok.

Things are different for the hiberimage LV. It spans several GB,
so we don't want to thicken it at boot time, which could result
in storage wear out. At boot time a dm-crypt device is set up
on top of the hiberimage LV. The DM device is locked down, which
means hiberman can't open it for thickening. Fortunately just
creating a dm-crypt device doesn't alter the data on the backing
device, so we can thicken the backing device, without impacting
the dm-crpyt device. The thickening is done just before creating
the snapshot image, to make sure the thinpool metadata of the
hibernated kernel matches the metadata on disk.

BUG=b:278146805
TEST=hibernate works on brya

Change-Id: I599dbe2b92e5206582433f3f00b8ab1867afdccc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4436012
Reviewed-by: Vovo Yang <vovoy@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>