shill: cellular: Do not fully reattach on Attach APN change
When the attach APN logic was added to shill a few years ago, the design
included a network reattach every time the APN changed. This was because
the logic was oversimplified and we needed to ensure that the modem
would always have the correct attach APN, and would not get stuck with a
bad attach APN. Since then, we've improved the logic by using multiple
attach APNs with round robin, adding a fallback APN and more
improvements. With this, the reattach is no longer needed.

BUG=b:217767378
TEST=Tested on multiple DUTs, run the entire suite on testhaus.

Cq-Depend:chromium:4515438
Change-Id: I26f8d6f887c5f7db6801853f506e0c3750dcfb29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4514840
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>