modemfwd: Run modem health ping in a sandbox
This patch generalizes modemfwd's sandbox implementation to be useful to
all required utilities, which now includes the modem helpers and modem
ping utilities for health monitoring.

The new implementation is more flexible and allows for sandboxed
processes to have a timeout enforced, which is useful in cases where a
process could potentially hang such as a frozen modem that no longer
response to pings.

This patch has been tested manually and via the tast.cellular.Modemfwd*
automated tests. Tested on hatch and brya to not regress helper and
health polling functionalities.

BUG=b:273304748
TEST=manual and automated flashing and recovery tests

Cq-Depend: chromium:4409875
Change-Id: I195eb0a8ad1976636c13dbd8375c8ffd5f32e2c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4426191
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>