missive: Start adding controlled degradation
This is the first stage: adding declarations and most of the code from
Chromium, but with no candidate queues provided for actual degradation
yet.

Note that in the face of the impending multi-generation support we can
no longer assume there is only one queue of each priority, and so the
degradation candidates queues cannot be constructed upfont - they need
to be produced anew every time we encounter low disk space condition.

For that purpose I isolate owner of storage queues into a separate
QueuesContainer class, but for now do not support enabling the
degradation. For that reason degradation code is added but not run yet (and not covered by the tests).

BUG=b:214038621
TEST=FEATURES=test emerge-${BOARD} missive

Change-Id: Ie079dca2d52c4720025cdb50d3a1c4982114b2a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4426637
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Josh Hilke <jrhilke@google.com>
Reviewed-by: Hong Xu <xuhong@google.com>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>