hiberman: Reduce hiberimage size to half of the RAM size
For hibernate the Linux kernel requires half of the system RAM to
be free as a temporary storage for the hibernate image, before it
is written to disk. As a result the hibernate image can't be
larger than half of the RAM size, so reduce the size of the
hiberimage volume to that. Also adjust the size of the
hiberintegrity volume accordingly.

While we are at it move the calculation of the number of pages
into the branch for the hiberintegrity volume, which is the only
place where it is needed.

BUG=b:278284140, b:278146805
TEST=hibernate works on brya

Change-Id: I51855689460d661d9655bb01bdd0b3990885cf67
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4436010
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Vovo Yang <vovoy@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>