shill: Passpoint pad 00s for OIs
Pad 00s to Passpoint OIs with unexpected length. OIs are expected to be
either OUI (24-bit) or OUI-36 (36-bit) of IEEE standards. This means
supplicant expects either a hex string of length 6 or 10.

The padding is necessary as ChromeOS OI value is taken from Android
which unfortunately does not have the necessary length field. Android
does the matching by converting both the client's and AP's OI values to
long.

This padding fix is not technically correct, but is the best possible
fix for ChromeOS given the missing OI length information. The matching
fails when an entity owning an OUI value of "000xxx" (hex
representation) creates their own OUI-36-based OI for Passpoint
"0000xxxyyy". When this happens, ChromeOS uses "xxxyyy" as their OI
value eventhough "0000xxxyyy" is needed.

BUG=b:277299855
TEST=FEATURES=test emerge-<BOARD> shill

Change-Id: I625a62c76a95452849a4c45af1e38ace5eb38ff7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4428631
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>