runtime_probe: Move allowed function lists to `all_functions.h`
The allowed function lists of different probe config loaders were
defined in the corresponding classes. When adding a probe function, it
is easy to miss adding it to the allowed list. This CL move the
definition of *ProbeConfigLoader::kAllowedProbeFunctionNames to
`all_functions.h` to prevent from this problem.

BUG=b:267594391
TEST=FEATURES=test emerge-corsola runtime_probe
TEST=Manually test on DUT

Change-Id: I38bf4425bf4937117b5a643facd6a8625334bf93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4432334
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Tested-by: Allen Shih <allenshihmc@chromium.org>
Reviewed-by: Kevin Lin <kevinptt@chromium.org>