diagnostics: Decouple cros-health-tool and libdiag_routine
The internal default values of routine parameters should not be
referenced directly by the clients. We can assign an arbitrary value to
the default value of the command line argument since it won't be used.
There should be a if-statement to check whether the argument is set or
not before passing the argument.

BUG=b:277703009
TEST=USE="-cros-debug" FEATURES=test emerge-volteer diagnostics

Change-Id: I46e6d7cdfbd2f8da4a2d72f28df50e563dea5840
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4431215
Tested-by: Wei-Luan Wang <weiluanwang@google.com>
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>
Reviewed-by: Byron Lee <byronlee@chromium.org>