missive: Fix the flaky test.
The test asynchronously posts an update callback and then expects
it to be called once parameters are changed. The posting could happen
too late, after the parameters change - and then the callback was not
invoked.

The fix adds a `done_cb` closure to the update callback posting, so that
the tests can make sure to change the parameters after that.

The flakiness was reproed with 100 runs of the test; after the fix the
same action did not hang up.

BUG=b:278469849
TEST=FEATURES=test emerge-${BOARD} missive (100 iterations)

Change-Id: I37a98ff1dc46d1961182ed3e2ec85741eb3681b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4435249
Reviewed-by: Hong Xu <xuhong@google.com>
Reviewed-by: Josh Hilke <jrhilke@google.com>
Auto-Submit: Leonid Baraz <lbaraz@chromium.org>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>