hiberman: Add helpers for volume creation
The new function VolumeManager::get_volume_properties() returns
properties of a logical volume, like its name, size and whether
the volume should be thickened at creation. This allows to remove
the 'name' parameter from create_or_activate_lv(), which is
somewhat redundant with the volume type parameter. The new
VolumeManager function create_lv() creates a thin volume and
thickens it if that's specified in the properties of the volume.

BUG=b:278146805
TEST=hibernate works on brya

Change-Id: I82b2254693b634c799c003d8fbe1407079896b43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4436011
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Vovo Yang <vovoy@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>