common-mk: Move brillo log settings to TestRunner
Currently, if a user customizes the call to `TestRunner`, this brillo
log setting is lost. It is more flexible to let `brillo::InitLog` be
called from `TestRunner`.  This is added as an option in case some edge
case requires turning this feature off.

Following up crrev.com/c/4343770

Note that brillo dependency is introduced to the header.

BUG=None
TEST=CQ

Change-Id: Ib3d4b68dea0416a77efa636bafcc9a0129e10be2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4436009
Tested-by: Hong Xu <xuhong@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Commit-Queue: Hong Xu <xuhong@google.com>