crash: Don't use "Chromium" as a product name in uploads.log
Remove the extra code that uses "Chromium" and "ChromiumOS" as product
names in uploads.log when on an unofficial build.

https://chromium.googlesource.com/chromium/src/+/main/styleguide/c++/c++.md#naming
says we shouldn't be using Chromium in the code anywhere, and this adds
complexity. In particular, removing this will make it easier to fix
crash.SenderDryLog tast test, because we won't need to replicate the
official / unofficial logic in the test.

Originally, the Chrome/Chromium product name distinction was make to
avoid contaiminating the crash dashboards with crashes from unofficial
builds, but (a) we have other ways to do that (unofficial builds will
only upload to the staging server) and (b) over the years, this code
changed to only affect uploads.log.

Also fixed a few other places that were in violation of the style guide.

BUG=b:278627471
Fixed: b:278627471
TEST=Uploaded crashes with '--dev', checked chrome://crashes, saw
"Chrome" and "ChromeOS"
TEST=tast -verbose run --extrauseflags chrome_internal my_crbook crash.SenderDryRun*

Cq-Depend: chromium:4437293
Change-Id: I96a2d2ffdc4358a4b0e61bec067923f23f44fda8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4437692
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Ian Barkley-Yeung <iby@chromium.org>