shill: Add UMA metric to track WiFi restart request
To mitigate the issue in b:270746800, we implemented a mechanism
for intel iwl7000 driver to send a uevent when it should be restarted.
This is not meant to be a permanent change, and we need a metric to
see how often this code path is hit. We add
Network.Shill.WiFi.RestartReason to track this statistic.

The plan is to remove this mechanism and metric once the hitrate is low
enough.

BUG=b:270746800
TEST=shill unit tests pass

Change-Id: I6b7bd71acf080283df3c5a6ade9ba3d0974ab9d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4443466
Commit-Queue: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>