ml_core: Add support for USE=local_ml_core_internal with ondevice_image_content_annotation
Both ml-core and ml-core-internal install the same destination files, depending on local_ml_core_internal.
Note that this should be safe for CQ as local_ml_core_internal is only for local development.
If this triggers file collisions then the easiest solution is to emerge -C ml-core ml-core-internal.
This will be addressed at some point, possibly by merging the ebuilds.

BUG=b:208909641, b:208909640
TEST=FEATURES=test USE="ondevice_image_content_annotation" emerge ml ml_core with and without USE=local_ml_core_internal

Change-Id: I34ee562cb2d4573c379b5b3952b4235777e7319a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4444809
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Tested-by: Nikolas Bowe <nbowe@google.com>