camera: Fix manipulator test segmentation issue
When a two consecutive frames are sufficently different, the segmentation
mask requires a sufficiently high iterations to blend from one mask to the
next. This causes the expected result to differ as it is still
transitioning to the expected result.

Updating sample image to the same image such that the test produce
correct result. This avoids us to use a very high number of iterations
slowing the test down.

BUG=b:278821053
TEST=cros_effects_sm_tests

Cq-Depend: chromium:4446197
Change-Id: Icff8a7c5342200df08484ad597027149d0ec77d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4446137
Tested-by: Jingyuan Tu <zactu@google.com>
Commit-Queue: Jingyuan Tu <zactu@google.com>
Reviewed-by: Jim Pollock <jmpollock@chromium.org>