cryptohome: Move the AuthFactor proto function to the drivers
This relocates GetAuthFactorProto to the driver code, where each factor
type can implement its own logic. The existing code was actually a
switch on top of a collection of type-specific functions so this is a
fairly natural refactor.

One catch is that there's also some common code to all of the types, so
I did a bit of refactoring to make that easier to reuse. Instead of just
having an interface -> concrete class structure, I added a middle layer
with a single common/helper class that partially implements the
interface with the common bits.

BUG=b:278937647
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I29fab616e777fba8bb4d0525b36d93a90b8b43f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4454749
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>