crash: Add ec_collect option to crash_reporter
In some cases the EC is able recover after a panic for a short period.
aka EC System Safe Mode. When this occurs the ec crash collector needs
to be triggered asynchronolsy.

This change adds a command line option to directly trigger the ec
collector. A udev rule is also added for trigger the ec collector from a
uevent.

BUG=b:258195448
TEST=Observe ec collector runs with crash_sender --ec_collect
TEST=Observe ec collector runs when uevent is triggered

Change-Id: Idf436e10db18f1b978d6042ac0ae657ccad283e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4450501
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>