hiberman: Use kernel key ring for dm-crypt keys
Currently the dm-crypt devices hiberimage and hiberintegrity are
set up with the encryption key as a hex string in the DM table.
This provides potential bad actors easy access to the keys
through 'dmsetup table'. Instead add the keys to the kernel
key ring, create the dm-crypt devices with references to these
keys and then remove the keys again from the kernel key ring.

BUG=b:279172349
TEST=hibernate works on brya
  keyctl list @s | grep hiber
    => nothing

Change-Id: I1740315a4b478acb3ec50cb65c7ee0fbea2d4396
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4459713
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>