runtime_probe: Fix |ProbeResultCheckerList| to validate correctly
The checkers convert the probe result in-place before the validation.
This may cause the checkers in the |ProbeResultCheckerList| to use the
probe result modified by the previous checker for validation, resulting
in false negative. This CL modifies the |ProbeResultCheckerList| to use
the copy of the probe result for conversion and validation, and only
update the probe result if the validation passes.

BUG=b:278995770
TEST=FEATURES=test emerge-corsola runtime_probe
TEST=Manually test list checkers with hex fields on DUT

Change-Id: Ibd414e9503377394b86fe529a71f0890ce366268
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4458408
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Tested-by: Allen Shih <allenshihmc@chromium.org>
Reviewed-by: Kevin Lin <kevinptt@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>