shill: net: add IPv4CIDR class
This CL adds the IPv4CIDR class to represent the CIDR for IPv4
protocol. Because the implementation is almost the same as the CIDR
for IPv6 protocol, we use template to implement the logic.

The IPv6CIDR will be added at the following CL.

BUG=b:269983153
TEST=FEATURES=test emerge-hatch shill-net

Change-Id: I481f21c2b9931e7e39116b7f855450bc58095448
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4458548
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>