chromeos-config: Add ARCVM global boost parameter
Add a parameter which can control the ARCVM vcores and services boost by
using the scaling factor. The new boost-arcvm parameter is part of the
scheduler-tune.

This is a way to improve the general ARCVM experience and to adjust the
boost value for slower and faster models from the same board family.
Please note, that this is intended to be a temporary solution, and will
be removed upon having the more appropriate scheduler tuning features
ready.

The boost factor can be applied directly in the machines with the same
CPU frequencies (1.0 means, the VM will be boosted by 100%, 0.0 means,
that there is no boost at all), but in order to give the higher boost to
the slower machines on the little.BIG architecture, the ratio of
little/BIG CPU frequency for scaling.

For example, take the slowest trogdor.lazor:
  - boost-arcvm: 0.82
  - little core max freq: 1.8 Ghz
  - big core max freq: 2.1 Ghz

By using this formula, the trogdor.lazor will be boosted by the 70%.

The faster trogdors, having the big core max frequency of 2.4 Ghz, will
be boosted by 62% with the same boost-arcvm value set to 0.82.

BUG=b:217825939
TEST=./regen.sh && ./run_tests.sh

Cq-Depend: chromium:4454321
Change-Id: I6a8f06507c6436c3ed9e0749a6738a2390491d41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4459614
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Marcin Wojtas <mwojtas@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>