vm_tools: concierge: Boost the ARCVM if boost-arcvm exists
If the boost-arcvm parameter was defined, it means, that the whole VM
should be boosted, by setting the vcores and host services cgroups
uclamp.min to a boost value which is scaled to a little/big cores
frequency ratio. This provides higher boost for a slower models.

The uclamp.latency_sensitive attribute is always set if the boost-arcvm
value is greater than 0.

BUG=b:243336790
BUG=b:254277702
TEST=emerge-strongbad vm_host_tools

Change-Id: I2fbd0a888770f00f799d51b4075b5214d59ebf6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4459615
Reviewed-by: Nic Hollingum <hollingum@google.com>
Tested-by: Marcin Wojtas <mwojtas@google.com>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>